Inline sigreturn() in handler_container
We shouldn't directly call sigreturn(), because we didn't map all the
kernel functions to user space. In the previous implementation, it
worked only because the distance between the handler and sigreturn was
less than 0x1000, which allowed the call to remain within a single
memory page. This reliance on proximity is fragile and could lead to
undefined behavior if the distance exceeds this limit or if the memory
layout changes.